+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added 
+### Added
+- Campaigns can be sorted by name in alphabetical order
 - "Reset filters" button clears filter options on the Campaigns page
 
 ### Fixed